[front] test: isolate model availability matrix
